requirements.txt
claude-agent-sdk>=0.1.48
lark-oapi>=1.5.0
mcp>=1.0.0
openai>=1.0.0
Pillow>=10.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
